Add support of new plugin/driver "vsphere-shared".
To build the shared plugin: make build-shared
Only docker-based build is supported for now.
Default config file path: /etc/docker-volume-shared.conf
Default log file path: /var/log/docker-volume-shared.log
