+++ client_plugin/default-config.json
