+++ client_plugin/drivers/vmdk/vmdkops/esx_vmdkcmd.go
