+++ client_plugin/drivers/vmdk/vmdkops/mock_vmdkcmd_unsupported.go
